requirements.txt
faiss-cpu
sentence-transformers
tqdm
huggingface-hub<1.0
